backend/app/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.23.2
httpx==0.24.1
cachetools==5.4.0
python-dotenv==1.0.0